requirements.txt
anthropic
openai
python-dotenv
fastapi
uvicorn
pydantic
pillow
httpx
aiofiles
prompt_toolkit
beautifulsoup4
requests
shazamio
sounddevice
soundfile
numpy
pypdf
python-docx
markdown
